requirements.txt
django==3.2.4
djangorestframework==3.12.4
psycopg2==2.8.6
prospector==1.3.1
factory-boy==3.2.0